Redefine genotype to remove context-related information from genome
Now the genome is back to an array of doubles ranging in [0, 1].
And it is phen's responsibility to convert those doubles to meaningful
values of the design vector. So with this setup, any arbitary array
with ENOUGH LENGTH represents a valid robot. Insufficient length is the
only occasion that makes a phen invalid.