Adjust the camera position according to the position of the environment
The height of the environment varies based on the height of the robot --
the root part of the robot is always placed at the origin (subject to
user change), and the environment is lowered to let the longest leg of
the robot appear above it. This casued some inconsistency in camera
views between different experiments.