Update descriptor assignment during training to tell different fate of the robot
Now the fit of phen has both descriptors initialized to -1 and fitness
to 0. If descriptors end up with -2, then the genome was not valid at
the beginning and were never developed to a robot.  And -3 means the
developed robot was self-colliding.

This differs from the case where the genome was valid, but the robot was
not added to archive, which made the grid id of both descriptors be -1